FEAT-#6351: Add a simple heuristic for fragment size when running on a GPU
